+++ lib/applitools/selenium/visual_grid/ios_device_info.rb
@@ -17,12 +17,11 @@ module Applitools
         end
       end
 
-
       def initialize(options = {})
         super()
         self.ios_device_info = EmulationInfo.new.tap do |ei|
           ei.device_name = options[:device_name]
-          ei.screen_orientation = options[:screen_orientation] || IosScreenshotOrientations::PORTRAIT
+          ei.screen_orientation = options[:screen_orientation] || options[:orientation] || IosScreenshotOrientation::PORTRAIT
         end
       end
 
@@ -30,7 +29,7 @@ module Applitools
 
       class EmulationInfo < EmulationBaseInfo
         enum_field :device_name, IosDeviceName.enum_values
-        enum_field :screen_orientation, IosScreenshotOrientations.enum_values
+        enum_field :screen_orientation, IosScreenshotOrientation.enum_values
 
         def json_data
           {